Minimize quality allowlist
Split oversized modules so only the CLAP ABI boundary remains allowlisted.
