+++ crates/wrac_clap_adapter/quality.toml
@@ -1,5 +1,4 @@
 [file_length]
 allow = [
     { path = "src/abi.rs", reason = "Existing CLAP ABI dispatch code keeps closely related raw callback wiring together. Splitting it should be handled separately with focused regression coverage." },
-    { path = "src/events.rs", reason = "Existing CLAP event conversion code keeps paired input and output event handling together. Splitting it should be handled separately with host behavior checks." },
 ]
