--- crates/wrac_wxp_gui/quality.toml
@@ -1,4 +1,0 @@
-[file_length]
-allow = [
-    { path = "src/controller.rs", reason = "Existing WebView controller lifecycle and platform integration are coupled in this file. Splitting it should be handled separately with UI smoke checks." },
-]
